requirements.txt
asgiref==3.2.3
cffi==1.14.0
cryptography==2.8
Django==3.0.3
django-crispy-forms==1.8.1
django-js-asset==1.2.2
django-mptt==0.11.0
django-rest-knox==4.1.0
django-taggit==1.2.0
djangorestframework==3.11.0
Pillow==7.0.0
pkg-resources==0.0.0
psycopg2==2.8.4
pycparser==2.20
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
